+++ test/OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests/PrometheusExporterMiddlewareTests.cs
@@ -281,7 +281,7 @@ namespace OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests
             }
             else
             {
-                Assert.Equal(HttpStatusCode.NoContent, response.StatusCode);
+                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
             }
 
             validateResponse?.Invoke(response);