+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol.Logs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* `OtlpExporterOptions` can now be bound to `IConfiguation` and
+  `HttpClientFactory` may be used to manage the `HttpClient` instance used when
+  `HttpProtobuf` is configured
+  ([#3640](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3640))
+
 ## 1.4.0-alpha.2
 
 Released 2022-Aug-18