+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol.Logs/CHANGELOG.md
@@ -9,7 +9,7 @@
 
 * Added overloads which accept a name to the `OpenTelemetryLoggerOptions`
   `AddOtlpExporter` extension to allow for more fine-grained options management
-  ([#XXXX](https://github.com/open-telemetry/opentelemetry-dotnet/pull/XXXX))
+  ([#3652](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3652))
 
 ## 1.4.0-alpha.2