+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol.Logs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased-Logs
+
+* Added overloads which accept a name to the `LoggerProviderBuilder`
+  `AddOtlpExporter` extension to allow for more fine-grained options
+  management
+  ([#XXXX](https://github.com/open-telemetry/opentelemetry-dotnet/pull/XXXX))
+
 ## Unreleased
 
 * `OtlpExporterOptions` can now be bound to `IConfiguation` and